requirements.txt
cffi==1.17.1
click==7.1.2
dlib==19.20.0
jpegtran-cffi==0.5.2
opencv-python==4.3.0.36
pycparser==2.22
